Enable ESLint
Although there was a minimal ESLint configuration before, it had no
rules enabled, and wasn't even checking the specs properly as it didn't
have modules enabled in there. This fixes that by bringing in the same
style guide used in almost all the other linter providers. This also
lists ESLint as a dependency, so checking against this should work in
the CI builds from now on.