Enable linting on change
It looks like we might be safe to move to using `stdin` input to
`clang` by changing the CWD to that of the file instead of the current
project/file setting. Since that was only a guess and hasn't went live,
we might as well start with this to see how things work out.